# src/lib/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Workers Webhooks API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Bearer token authentication"
      },
      "apiKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "apiKeyCookie",
        "description": "API Key authentication via cookie"
      }
    },
    "responses": {
      "ErrorResponse": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error message"
                }
              }
            },
            "example": {
              "error": "Internal server error"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyCookie": []
    }
  ],
...
  "paths": {
...
    "/event-types": {
      "get": {
        "summary": "List Event Types",
        "description": "List event types for the current environment",
        "tags": [
          "Event Types"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "eventTypes": {
                      "type": "array",
                      "description": "List of event types",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "Event type ID"
                          },
                          "environmentId": {
                            "type": "string",
                            "description": "Environment ID"
                          },
                          "name": {
                            "type": "string",
                            "description": "Event type name"
                          },
                          "description": {
                            "type": "string",
                            "description": "Event type description"
                          },
                          "schema": {
                            "type": "object",
                            "description": "JSON Schema for event payload validation (must be valid JSON Schema)"
                          },
                          "enabled": {
                            "type": "boolean",
                            "description": "Whether the event type is enabled"
                          },
                          "createdAt": {
                            "type": "string",
                            "description": "Created at"
                          },
                          "updatedAt": {
                            "type": "string",
                            "description": "Updated at"
                          }
                        },
                        "example": {
                          "id": "user.created",
                          "environmentId": "a1b2",
                          "name": "User Created",
                          "description": "User created event",
                          "schema": {
                            "type": "object",
                            "properties": {
                              "userId": {
                                "type": "string"
                              }
                            },
                            "required": [
                              "userId"
                            ]
                          },
                          "enabled": true,
                          "createdAt": "2021-01-01T00:00:00.000Z",
                          "updatedAt": "2021-01-01T00:00:00.000Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Due to missing or invalid authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. You do not have permission to access this resource or to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests. You have exceeded the rate limit. Try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Rate limit exceeded"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. This is a problem with the server that you cannot fix.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "x-speakeasy-group": "eventTypes",
        "x-speakeasy-name-override": "list"
      },
      "post": {
        "summary": "Create Event Type",
        "description": "Create new event type",
        "tags": [
          "Event Types"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Name of the event type",
                    "example": "User Created"
                  },
                  "description": {
                    "type": "string",
                    "description": "Description of the event type",
                    "example": "User created event"
                  },
                  "schema": {
                    "type": "object",
                    "description": "JSON Schema for event payload validation (must be valid JSON Schema)",
                    "example": "{\n  \"type\": \"object\",\n  \"properties\": {\n    \"userId\": {\n      \"type\": \"string\"\n    }\n  },\n  \"required\": [\"userId\"]\n}\n"
                  },
                  "enabled": {
                    "type": "boolean",
                    "description": "Whether the event type is enabled",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "Event type ID"
                    },
                    "environmentId": {
                      "type": "string",
                      "description": "Environment ID"
                    },
                    "name": {
                      "type": "string",
                      "description": "Event type name"
                    },
                    "description": {
                      "type": "string",
                      "description": "Event type description"
                    },
                    "schema": {
                      "type": "object",
                      "description": "JSON Schema for event payload validation (must be valid JSON Schema)"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Usually due to missing parameters, or invalid parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Invalid schema"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Due to missing or invalid authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Unauthorized"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. You do not have permission to access this resource or to perform this action.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Forbidden"
                  }
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests. You have exceeded the rate limit. Try again later.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Rate limit exceeded"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. This is a problem with the server that you cannot fix.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/responses/ErrorResponse",
                  "example": {
                    "error": "Internal server error"
                  }
                }
              }
            }
          }
        },
        "x-speakeasy-group": "eventTypes",
        "x-speakeasy-name-override": "create"
      }
    },
...
  }
}